mediacapture-depth/MediaStreamTrack-end-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test that mediastreamtrack are properly ended</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#life-cycle">
<link rel="help" href="https://w3c.github.io/mediacapture-depth/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<p class="instructions">Use a test device with depth camera(embedded or external).</p>
<p class="instructions">When prompted, accept to share your depth stream, and then revoke that permission.</p>

<div id="log"></div>

<script>
// longer timeout since requires user interaction
setup({ explicit_timeout: true });
async_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    let vidTrack = stream.getVideoTracks()[0];
    vidTrack.onended = t.step_func_done(() => {
      assert_equals(vidTrack.readyState, "ended", "Video track has been ended as expected");
      assert_false(stream.active, "MediaStream has been inactive as expected");
    });
    assert_equals(vidTrack.readyState, "live", "The video track object is in live state");
  });
}, "Tests that the video MediaStreamTrack objects are properly ended on permission revocation");

</script>
